OOSD Assignment 5 updated

+++ CS5700 - Object Oriented Software Development/Assignment 5/Assignment 5/src/com/usu/sudokuSolver/template/SudokuSolver.java
@@ -96,7 +96,7 @@ public abstract class SudokuSolver {
 	}
 	
 	private static boolean validBoardSize(String board[][]) {
-		if(!boardSizes.contains((int)Math.sqrt(board.length))) {
+		if(!boardSizes.contains(Math.sqrt(board.length))) {
 			return false;
 		}
